Refactoring

+++ com/BombingGames/Game/Chunk.java
@@ -173,8 +173,8 @@ public class Chunk {
                             }
 
                             data[x][y][z] = new Block(
-                                        (int) Integer.parseInt(line.substring(0,posdots)),
-                                        (int) Integer.parseInt(line.substring(posdots+1, posend))
+                                        Integer.parseInt(line.substring(0,posdots)),
+                                        Integer.parseInt(line.substring(posdots+1, posend))
                                         );
                             x++;
                             line.delete(0,posend+1);
